Refactor user profile update logic and enhance DTO consistency
Revised the logic for updating user profiles in the UsersService and UsersController, ensuring cleaner and more consistent functionality. Simplified DTO validation using custom decorators and updated the method to check for authorized users. Refactored redundant code in user-related modules for improved readability and maintainability.
