+++ src/modules/users/users.controller.ts
@@ -1,9 +1,12 @@
-import { Controller, Get, Query, ValidationPipe } from '@nestjs/common';
+import { Body, Controller, ForbiddenException, Get, Patch, Query } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { Auth, UUIDParam } from '../../decoractors/http.decorators';
 import { RoleType } from '../../constants/role-type';
 import { UsersPageOptionsDto } from './dto/user-page-options.dto';
 import type { UserDto } from './dto/user.dto';
+import { AuthUser } from '../../decoractors/auth-user.decorators';
+import { UserEntity } from './entities/user.entity';
+import { UpdateUserProfileDto } from './dto/update-user-profile.dto';
 
 @Controller('api/users')
 export class UsersController {
@@ -23,6 +26,20 @@ export class UsersController {
 		return this.usersService.getUser(userId);
 	}
 
+	@Patch(':id')
+	@Auth([RoleType.USER])
+	async updateProfile(
+		@AuthUser() user: UserEntity,
+		@UUIDParam('id') userId: Uuid,
+		@Body() updateProfileDto: UpdateUserProfileDto
+	){
+		if (userId !== user.id){
+			throw new ForbiddenException('You can only update your own profile')
+		}
+
+		return this.usersService.editUserProfile(user, updateProfileDto);
+	}
+
 	// @UseGuards(JwtAuthGuard, OwnerOrAdminGuard)
 	// @Get(':id')
 	// async getUserById(@Param('id') id: string, @Req() req) {
@@ -38,16 +55,4 @@ export class UsersController {
 	// 	return this.usersService.unbanUser(id);
 	// }
 
-	// @Get('profile/:userId')
-	// async getUserProfile(@UUIDParam('userId') userId: Uuid) {
-	// 	return this.usersService.getUserProfile(userId);
-	// }
-
-	// @Put('profile/:userId')
-	// async editProfile(
-	// 	@UUIDParam('userId') userId: Uuid,
-	// 	@Body() updateData: UpdateUserProfileDto,
-	// ) {
-	// 	return this.usersService.editUserProfile(userId, updateData);
-	// }
 }
